refactor: ssr loadable-component collector
